unreuse user api tests
