+++ vault-plugins/flant_iam/model/tenant.go
@@ -13,6 +13,7 @@ const (
 type Tenant struct {
 	UUID       string `json:"uuid"` // ID
 	Identifier string `json:"identifier"`
+	Version    string `json:"resource_version"`
 	// TODO enabled_by_default_for_new_projects
 	// TODO resource_version
 }
